Finished pretty much the core functionalities w/out the report generation. Need to add Login and Books API, with the report generation.

+++ Assignments/Assignment 2/src/main/java/bookStore/controller/EmployeeFormController.java
@@ -6,8 +6,11 @@ import bookStore.service.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.ModelAndView;
 
+import javax.validation.Valid;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -23,27 +26,27 @@ public class EmployeeFormController {
     }
 
     @GetMapping("/employeeCreateForm")
-    public String creatingForm(Model model) {
-        EmployeeDto employeeDto = new EmployeeDto();
-        model.addAttribute("employee", employeeDto);
+    public String creatingForm(EmployeeDto employeeDto) {
         return "/employeeCreateForm";
     }
 
     @PostMapping("/employeeCreateForm")
-    public String greetingSubmit(@ModelAttribute EmployeeDto employeeDto) {
+    public String greetingSubmit(@Valid EmployeeDto employeeDto, BindingResult bindingResult) {
+        if (bindingResult.hasErrors())
+            return "/employeeCreateForm";
         employeeService.create(employeeDto);
         return "/success";
     }
 
     @GetMapping("/employeeDeleteForm")
-    public String deleteForm(Model model){
-        EmployeeDto employeeDto = new EmployeeDto();
-        model.addAttribute("employee", employeeDto);
+    public String deleteForm(EmployeeDto employeeDto){
         return "/employeeDeleteForm";
     }
 
     @PostMapping("/employeeDeleteForm")
-    public String deleteEmployee(@ModelAttribute EmployeeDto employeeDto) {
+    public String deleteEmployee(@Valid EmployeeDto employeeDto, BindingResult bindingResult) {
+        if (bindingResult.hasFieldErrors("username"))
+            return "/employeeDeleteForm";
         employeeService.delete(employeeDto);
         return "/success";
     }
@@ -58,6 +61,19 @@ public class EmployeeFormController {
         EmployeeDto employeeDto = employeeService.findByUsername(fields.get("username"));
         employeeService.update(employeeDto, fields.get("newUsername"));
         return "/success";
+    }
+
+    @GetMapping("/employeeFindForm")
+    public String findForm (Model model){
+        return "/employeeFindForm";
+    }
+
+    @PostMapping("/employeeFindForm")
+    public ModelAndView findEmployee(@RequestParam Map<String, String> field){
+        EmployeeDto employeeDto = employeeService.findByUsername(field.get("username"));
+        ModelAndView mav = new ModelAndView("employeeFound");
+        mav.addObject("employee", employeeDto);
+        return mav;
     }
 
 }
